chore: cleanup & notes
