+++ src/juju_doctor/probes.py
@@ -2,9 +2,7 @@
 
 import importlib.util
 import inspect
-import json
 import logging
-from collections import defaultdict
 from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Dict, List, Optional
@@ -15,12 +13,15 @@ import fsspec
 import yaml
 from rich.console import Console
 from rich.logging import RichHandler
-from treelib.tree import Tree
 
 from juju_doctor.artifacts import Artifacts
 from juju_doctor.fetcher import FileExtensions, copy_probes, parse_terraform_notation
 
 SUPPORTED_PROBE_TYPES = ["status", "bundle", "show_unit"]
+EMOJI_MAP = {
+    "green": "🟢",
+    "red": "🔴",
+}
 
 logging.basicConfig(level=logging.WARN, handlers=[RichHandler()])
 log = logging.getLogger(__name__)
@@ -40,6 +41,15 @@ def _read_file(filename: Path) -> Optional[Dict]:
     except Exception as e:
         log.warning(f"Unexpected error while reading '{filename}': {e}")
     return None
+
+
+@dataclass
+class OutputFormat:
+    """Track the output format for the application."""
+    verbose: bool
+    format: Optional[str]
+    grouping: List[str]
+    exception_logging: bool
 
 
 @dataclass
@@ -172,15 +182,6 @@ class Probe:
 
 
 @dataclass
-class OutputFormat:
-    """Track the output format for the application."""
-    verbose: bool
-    format: Optional[str]
-    grouping: List[str]
-    exception_logging: bool
-
-
-@dataclass
 class ProbeResults:
     """A helper class to wrap results for a Probe."""
 
@@ -207,8 +208,8 @@ class ProbeResults:
 
     def text(self, output_fmt: OutputFormat) -> str:
         """Probe results (formatted as Pretty-print) as a string."""
-        green = EmojiTree.green
-        red = EmojiTree.red
+        green = EMOJI_MAP["green"]
+        red = EMOJI_MAP["red"]
         if self.passed:
             return f"{green} {self.probe_name} passed"
         # If the probe failed
@@ -251,7 +252,7 @@ class RuleSet:
             match ruleset_probe["type"]:
                 # TODO We currently do not handle file extension validation.
                 #   i.e. we trust an author to put a ruleset if they specify type: ruleset
-                case "directory" | "scriptlet":  # TODO Support a dir type since UX feels weird without?
+                case "scriptlet":
                     probes.extend(
                         Probe.from_url(ruleset_probe["url"], self.probe.probes_root, self.probe.get_chain())
                     )
@@ -281,113 +282,3 @@ class RuleSet:
                     raise NotImplementedError
 
         return probes
-
-
-class EmojiTree(Tree):
-    """A subclass of treelib.Tree that renders emoji shortcodes."""
-    red = ":red_circle:"
-    green = ":green_circle:"
-    EMOJI_MAP = {
-        green: "🟢",
-        red: "🔴",
-    }
-
-    def show(self, *args, **kwargs):
-        """Overrides show to replace emoji shortcodes with actual emojis."""
-        output = super().show(*args, stdout=False)  # Get tree output as string
-        if output:
-            for shortcode, emoji in self.EMOJI_MAP.items():
-                output = output.replace(shortcode, emoji)  # Replace shortcodes with emojis
-        else:
-            output = "Error: No tree output available."
-        if kwargs.get("stdout", True):
-            console.print(output)
-
-
-        output = super().show(*args, stdout=False)  # Get tree output as string
-
-        # Check if output is None, and handle the case if it is
-        if output is not None:
-            for shortcode, emoji in self.EMOJI_MAP.items():
-                output = output.replace(shortcode, emoji)  # Replace shortcodes with emojis
-        else:
-            # Handle the case where output is None, maybe log an error or set a default value
-            output = "Error: No tree output available."
-
-        if kwargs.get("stdout", False):
-            console.print(output)
-
-
-
-class ProbeResultAggregator:
-    """Aggregates and groups probe results based on metadata."""
-
-    groups = ["status", "artifact", "directory"]
-
-    def __init__(self, results: List[ProbeResults], output_fmt: OutputFormat):
-        """Prepare the aggregated results and its tree representation."""
-        self.results = results
-        self.output_fmt = output_fmt
-        self.tree = EmojiTree()
-        self.tree.create_node("Results", "root")  # root node
-        self.grouped_by_status = defaultdict(list)
-        self.grouped_by_artifact = defaultdict(list)
-        self.grouped_by_directory = defaultdict(list)
-        self._group_results()
-
-    def _group_results(self):
-        """Groups results by status, parent directory, and probe type."""
-        for result in self.results:
-            self.grouped_by_status[result.status].append(result)
-            self.grouped_by_artifact[result.func_name].append(result)
-            self.grouped_by_directory[result.probe.path.parent].append(result)
-
-    def _get_by_status(self, status: str) -> List[Dict]:
-        return self.grouped_by_status.get(status, [])
-
-    def _get_by_artifact(self, probe_type: str) -> List[Dict]:
-        return self.grouped_by_artifact.get(probe_type, [])
-
-    def _get_by_parent(self, parent: Path) -> List[Dict]:
-        return self.grouped_by_directory.get(parent, [])
-
-    def _build_tree(self, group: str) -> Tree:
-        tree = Tree()
-        tree.create_node(group.capitalize(), group)  # sub-tree root node
-        grouped_attr = getattr(self, f"grouped_by_{group}")
-        for key, values in grouped_attr.items():
-            tree.create_node(str(key), f"{group}-{key}", parent=group)
-            for probe_result in values:
-                tree.create_node(
-                    probe_result.text(self.output_fmt),
-                    f"{group}|{probe_result.func_name}" + probe_result.probe.get_chain(),
-                    parent=f"{group}-{key}",
-                )
-        return tree
-
-    def assemble_trees(self):
-        """For each group, build a sub-tree which gets pasted to the root tree of results.
-
-        Optionally display the exception logs and the tree result.
-        """
-        for group in self.output_fmt.grouping:
-            tree = self._build_tree(group)
-            self.output_fmt.exception_logging = False  # only log Exceptions once when grouping
-            self.tree.paste("root", tree)  # inserts the grouping's subtree into the aggregated tree
-
-    def print_results(self):
-        """Handle the formating and logging of probe results."""
-        total_passed = len(self._get_by_status("pass"))
-        total_failed = len(self._get_by_status("fail"))
-        match self.output_fmt.format:
-            case None:
-                self.assemble_trees()
-                self.tree.show(line_type="ascii-exr")
-                console.print(f"\nTotal: 🟢 {total_passed} 🔴 {total_failed}")
-            case "json":
-                self.assemble_trees()
-                tree_json = json.loads(self.tree.to_json())
-                tree_json.update({"passed": total_passed, "failed": total_failed})
-                print(json.dumps(tree_json))
-            case _:
-                raise NotImplementedError
